things are more broken somehow jc
plan: each type of attack will be a separate scene. bulletspawners will be all bullets, make it so that the init actually works (process false in ready so that process is not called before init). music attack controllers will contain tscn of all their attacks for that enemy. also: make bullets generic, and then have a animatedsprite tscn to add to it via the bulletspawner or musicAttackController?